requirements.txt
python-telegram-bot
python-dotenv
pymongo
motor 
aiohttp
requests
